refactor(ui): 移动并重构设置相关工具函数至utils目录
- 将App.tsx中的设置读写及相关辅助函数移除
- 在ui/utils/index.ts新增对应工具函数实现，包括读取写入用户及项目设置
- 将DEFAULT_MODEL和DEFAULT_BASE_URL常量移至constants.ts统一管理
- 优化导入路径，修正组件和工具的引用路径
- 修改RawModeContext和openai-client模块中依赖import路径
- 统一和调整DropdownMenu相关组件的导入路径
- 使代码结构更清晰，职责划分更明确，提高维护性
