+++ src/ui/components/index.ts
@@ -4,3 +4,4 @@ export { RawModeExitPrompt } from "./RawModeExitPrompt";
 export { default as SkillsDropdown } from "./SkillsDropdown";
 export { default as ModelsDropdown } from "./ModelsDropdown";
 export { default as FileMentionMenu } from "./FileMentionMenu";
+export { default as DropdownMenu } from "./DropdownMenu";
